testbot/VMs: Put the default type and status values first.
This way these will be the default in both MySQL and the GUI.  New VMs
are usually added as 'extra' and get promoted to 'base' when they have
been tested. They should also initially be in the 'dirty' state so
WineTestBot knows to put them it a clean state before using them.