testbot/bin: Make sure WineSendLog does not spam anyone by default.
This adds a $WinePatchToOverride configuration option and sets it to
the WineTestBot administrator email address by default. So as the name
implies any test result will be sent to him instead of to the patch
authors until he decides the WineTestBot to be worthy.  This also adds
a $WinePatchCc configuration option to set who should be CC-ed the
results. By default this is empty.