testbot/bin: Make the wine-patches script names more descriptive and consistent.
Prefix them with 'WinePatches' in reference to both the Wine project
and the wine-patches they relate to.  The 'WinePatchesML' script deals
directly with the wine-patches Mailing List . The Submit script
submits the patch directly.  The 'WinePatchesWeb' scripts deal with
Wine's Patch web site. Specifically the Notify script informs the
WineTestBot when new patches are available and the Get script grabs
and submits them for testing.