testbot: Fix the Job and Step status update.
A job that has a non-build step fail should still be marked as
failed. That's because when there are test failures the TestFailures
field gets set to a non-zero value but the task's Status is still set
to completed. So a failed Step really reflects a bot error.  Also
simplify the code and avoid duplicating the UpdateStatus() logic in
the Check() method.