testbot/TestAgent: Change the protocol and API for increased flexibility.
The old protocol relied on being able to close the sending side of the
connection to signal the end of the request, and then iusing the other
side to read the reply. However this is not always feasible,
especially when adding support for tunneling. Also it makes it
impossible to use a single connection for multiple requests.  In the
new protocol each request is composed of an RPC id, an argument count,
and for each argument its type, size and then data. So the other side
always knows exactly how much to read and thus a given connection can
be used for multiple RPCs.  The Perl TestAgent API is now
object-oriented which makes connection reuse simple.  The set of
operations provided by the Perl TestAgent module has also been changed
to be more flexible. In particular, where RunScript() was duplicating
the functionality of SendFile(), we now have Run(), Wait() and Rm()
which each perform a single function.  This change also unburdens the
TestAgent server from the selection of temporary filenames, and
eliminates a hard-coded path.  The internal client and server APIs
have also been tweaked to make sending and receiving the RPCs and
their replies easier.