testbot: Add a field to store the VM configuration details.
The description field is shown next to the VM name and thus has to be
reasonably short.  The details field is here so developers know
exactly what's special about the VM so they can figure out why their
test does not work.  Send the VM configuration details with the
WineTest report.