testbot/web: Improve the job status reporting on the home page.
It is interesting to know if there were test failures even if a VM
misbehaved. So report test failures even if some of the job's tasks
failed.  Conversely, don't claim that jobs like the 'Wine update' have
0 test failures.  Colorize the status to make errors and test failures
easier to spot.  Also link to the job details page from the status
field.