testbot/TestAgent: Extend and fix the run RPC's redirection options.
Always set up the stdout and stderr redirections for append so they
don't overwrite each other.  Add options to prevent the truncation of
the files stdout and stderr are redirected to. This makes it possible
to run multiple commands in a row and to collect all their output in a
single file.