testbot/VMs: Fix counting the powered on non-base VMs.
VMs with a 'disabling' role must be ignored.
Make a clear distinction between the 'enabled' VMs which we could use
but may currently be powered off (e.g. dirty VMs), and the 'powered
on' ones that use resources right now.  Also make it clear that these
methods are based on the Status and Role fields, not on what the VM
hypervisor thinks.