testbot/web: Make JobDetails.pl handle empty err logs correctly.
The task scripts used to create the err log and then delete it if it
was empty. However the Task module now redirects the script's stderr
to the err log to help diagnose script bugs (e.g. perl errors). That
means the script must only append to the err log and should certainly
not delete it.  So JobDetails.pl should now be able to deal with empty
err logs.