testbot/Users: Fix the deletion of user accounts.
Users cannot be deleted directly because the Jobs table hold
references to them. So mark the accounts as deleted and let the
Janitor.pl script delete them when they are not referenced anymore.
To do so, transform the 'Active' boolean field into a 'Status'
enumeration that can be 'active' (previously Active=Y),
'disabled' (Active=N) or 'deleted'.  The Sessions table also contains
references to the Users, but these should be deleted together with the
user account, which will forcefully log out the user as a side effect.