testbot: Rework the job scheduler and add an 'off' VM state.
The old scheduler would limit the number of VMs in the running state to $MaxRunningVMs but would allow dirty base VMs to accumulate without bound. The new scheduler makes sure that no more than $MaxActiveVMs VMs are using memory, CPU or I/O resources, whether they are reverting, sleeping, running or dirty.

Furthermore the new scheduler will power off any dirty VM that it cannot revert right away to make sure it won't waste resources.
The run tasks also no longer power off the VMs when done. They now leave that task to the scheduler which will decide what to do with the VM.
The new scheduler can revert VMs while others are running. This case gets its own limit, $MaxRevertWhileRunningVMs, so the reverts do not disturb the tests.
To better exploit this it can look ahead and prepare the VMs needed for a job's next steps. And like the previous scheduler, when idle it will start a set of VMs for the next job, trying to pick the most likely to be needed: first the build VM, then base VMs, etc.
This can also gets its own limit, $MaxVMsWhenIdle, so developpers can set the TestBot to not use resources unless needed.
This can lead to having idle VMs that turn out to not be needed, for instance if the new job targets non base VMs. When that happens the scheduler can shut down some idle VMs to make room for more useful ones.