testbot/lib: Make sure we have a consistent view of the VMs Status, no matter which Task we get the corresponding object from.
This ensures we don't schedule two tasks on the same VM when there is more than one VM host.

Specifically, in the single VM host scenario the scheduler actions go
like this:

  Schedule Jobs on Host1
    Load Job1 from database
      Load Step1 from database
        Load Task1 from database
          Load VM1 (e.g. buildvm) from database
          VM1 is idle -> start Task1
          VM1->Status = running
          Save VM1 to the database
    Load Job2 from database
      Load Step2 from database
        Load Task2 from database
          Load VM2 (e.g. the same buildvm) from database
           -> VM2->Status is up to date
          VM2->Status is running -> nothing to do
  All done

But in the two (or more) VM hosts scenario the first pass is when all
the objects are being loaded from database, while the second one just
reuses the already created Collections and Items. So in some cases
scheduling goes like this:

 Schedule Jobs on Host1
   Load Job1 from database
     Load Step1 from database
       Load Task1 from database
         Load VM1 (e.g. buildvm) from database
         VM1 is not on Host1 so continue
   Load Job2 from database
     Load Step2 from database
       Load Task2 from database
         Load VM2 (e.g. the same buildvm) from database
         VM2 is not on Host1 so continue
 Schedule Jobs on Host2
   Job1 is already in memory
     Step1 is already in memory
       Task1 is already in memory
         VM1 is already in memory
         VM1 is idle -> start Task1
         VM1->Status = running
         Save VM1 to the database
   Job2 is already in memory
     Step2 is already in memory
       Task2 is already in memory
         VM2 is already in memory
         -> VM2->Status has not been updated so
         VM2->Status is still idle -> start Task2
         -> Bug!!!

At this point we have two tasks trying to run on a single VM. When the
first one completes the VM gets powered off, leaving the second one
hanging.

With this patch VM1 and VM2 are the same object, and thus we
always have VM1->Status == VM2->Status which fixes this issue.