testbot/WinePatchesMLSubmit: Better handle errors and times when the Engine is not running.
WinePatchesMLSubmit must return a non-zero exit code in case it could
not deliver the message so the caller gets a chance to handle the
situation gracefully.  Have the Engine scan the staging directory for
pending patches from the mailing list and automatically handle them,
in particular on startup. This makes it possible for
WinePatchesMLSubmit to deliver the patches even if the Engine is not
running which is more robust.  This also fixes a potential source of
staging files leak.