testbot/web: Use the default Domain and Path of our cookies.
The Path defaults to '/' which is what we want.  The Domain is set to
the most restrictive value which is good.  This also means we're no
longer putting the port number in the Domain value which fixes our
cookies when the web server is running on a non-standard port or the
connection is going through port forwarding.