testbot/lib: Improve the PropertyDescriptor methods argument handling.
This standardizes the way we handle the arguments.  Make it so one can
readily see the list of arguments the method takes without having to
hunt down the implementation of all its ancestors.  Because these are
methods the added prototypes have no effect whatsoever as far as Perl
is concerned. However they do document which arguments are optional.