testbot/testagentd: Add a RemoveChildProc RPC.
The Wait RPC no longer causes the server to forget about the child
process. That reponsibility is transferred to the RemoveChildProc
RPC. This is to solve the case where the network connection breaks at
just the time when the child process terminates which could cause the
server to forget the child process exit status despite it not having
reached the client process.