testbot: Add proper command line processing and --debug and --help options to the Task and VM scripts.
The --debug option disables the VM status checks. This makes it
possible to put a VM in maintenance mode and then manually run the
scripts against it, without interference from the Engine.  This also
makes the VM status checks cleaner.  The --debug option also causes
log messages to be duplicated to stdout and additional traces to be
printed so one can follow the script's progress. Concise timing
information is also provided which makes it possible to figure out
where the script spends its time.