winegecko.php: Make use of autoselect mirror SourceForge feature.
This leaves the job of redirecting to mirror to SourceForge, which may
do much better job at this than we do. Our current solution is to
redirect to random mirror known to have Gecko. In our current code,
connection time is bad (it may even feel like it hangs), because we need
an additional request from server script to a mirror to ensure that the
resource is present there. Also, choosing random mirror often causes
pretty bad transfers.