testbot/web: Fix the sort order of a job's tasks.
Each entry's id is a numeric value that's equal to step.no*100+task.no.
But by default sort performs a string comparison resulting in task
(10,1), aka "1001", coming before (1,1), aka "101", which comes before
(11,1), aka "1101", etc.

Also note that sorting is normally the job of the CollectionBlock but
the JobDetails page is bypassing it.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>