testbot/web: Take into account the optional pid when reporting timeouts.
The 'done' line will not have a test unit name if there was no command
line parameter. Furthermore it may or may not have a process pid too.
So make sure that when the JobDetails page shows a timeout error it does
not mistake the pid for the test unit name.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>