winetest/gather: Add the test unit and group or tag to the tooltips.
The platform result pages can have quite a lot of columns so it is
easy to lose track of the test unit a line corresponds to.
Similarly the index pages have a lot of lines so it is easy to lose
track of which group or tag a column corresponds to.
So add this information to the cell tooltips.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>