winetest/gather: Remove support for the 'winetest crash' case.
dissect has no support for writing this type of status in the
summary.txt files.
Also if WineTest does crash one would likely have to manually submit
the results (i.e. it's unlikely to happen).
And most importantly a lot of test results would be missing which,
based on the current policy, means the whole report is and should be
rejected.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>