winetest/gather: Show the short date in the index titles and headers.
Since there is only one build per day, using the short date is just as
unique while being more readable and shorter. It has the additional
advantage of being ordered unlike the build id (*) which makes the
browser tab names more meaningful.
The build id is still shown in the page header in addition to the short
date since there is more space there and it is used for linking to
other pages.

(*) 'Mar 15' is clearly older than 'Mar 17' whereas no such
    determination can be made from '6a05069a' and 'c3b108e5' alone.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>