winetest/dissect: Use more informative titles for the report pages.
So far opening multiple pages would result in a bunch of brower tabs
all called 'Wine test run report', making it hard to keep track of
which is which.
Now they will have the short date, test unit if appropriate, and report
tag. So no matter if you try to figure out when the output of a
specific test changed, or want to compare the results of a given test
on multiple machines, or simply work on a few tests at once, you'll
know which tab contains the data you need.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>