testbot/WineRunTask: Tweak matching for the start and done lines.
Allow dll names to contain a dash as we have many such dlls in Wine.
Allow garbage at the start of the done line, since we can use the
current dll name to identify where it starts. But don't allow garbage at
the beginning of the start line as it could interfere with the proper
detection of the dll name.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>