testbot: Rebuild TestLauncher as part of the Reconfig.
Currently the TestLauncher binaries are checked into Git which is wrong
and means they are effectively never updated. Now that Reconfig.pl rebuilds
them the TestBot server will be able to retrive the up-to-date binaries
and use them.
Modify the TestLauncher makefile so it does not overwrite the checked-in
TestLauncher binaries.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>