testbot/WineRunTask: Better deal with single test unit executables.
If a Wine test executable has a single test unit, then one can upload
and run it without a parameter. In that case TestLauncher does not the
test unit name and thus does not print it on the start line. So
WineRunTask ends up with $CurrentUnit="" and must account for it when
it compares it to the test unit of skip lines.
This should also be factored when determining whether the test output
follows the Wine test standards.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>