testbot: Tweak the scripts so they print error messages on stderr by default.
When the scripts are started by the Engine their stderr is redirected to
the log in order to also capture Perl errors.
Since the scripts always log error messages, to avoid getting duplicates
they were only printing errors to stderr if given the --debug option.
As a result usage errors were not printed by default.
So the scripts now print the messages to stderr by default and the
Engine can invoke them with the --log-only option where needed to avoid
the stderr duplicates.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>